feat :CGIパーサーとの統一化のために抽象クラスの実装をした。ただ、抽象クラス自体がほぼ全てhttp_request_parserのためのものになってしまっているので、なんか微妙かも
